Removed domain-specific speedup.
There is still a modest speed increase to be gained by computing
feedForward() in a domain-specific way, but for now we let it go
and are content with the vectorized full calculation.  The difference
is about 15 percent on my current machine.

For now, we prefer not to have two versions in the codebase, because
we will be adding back learning and it is easier to have just one.  Once
that happens, we might consider putting it back.