Eliminated need for transpose.
We use a row vector instead of a column vector for the second
parameter weights, to avoid having to transpose to use matrix
multiplication.  The speedup is only slight because the compiler was
already figuring this out, but the code is more mathematically
informed.

We also removed the unneeded Transpose() and similar factory
methods from matrix, since we do not intend to use them and
we want to keep the matrix class very small.