Switching to full squash() calculation.
The fast version of squash() seems to introduce numerical stability
issues during training, probably because it has zero derivative on
the tails.

The speed difference is significant, but we will just
have to live with it for training.

We also plan to investigate more modern activation functions and
perhaps move away from squash() altogether.