Continuing rearranging temporal difference code.
Note that we have effectively changed the scale of the learning rate
hyperparameter "alpha" to be 36x what it was before.  I think it was
basically incorrect before, but relatively harmless because you could
set alpha to be 36x larger than it would have been otherwise.

The gradient calculation now computes dy/dx correctly (we think), where
y is the equity output of the net, and x is any parameter.  Previously,
we were computing, effectively (1/6) dy/dx and also dividing the error
term by 1/6, for a 1/36 total factor.

It does not change what the nets actually can learn, but it does change
what alpha you end up using, and is important to keep in mind when
comparing with training schedules from the past.

The intent of all this rework is to separate the temporal difference
calculation from the lowest-level neural net, which will then let
us try totally-different network structures, activation functions,
etc., without having to rewrite as much code.

This process is maybe halfway complete presently.  In the end, all
of the explicit temporal difference code should move out of net.h
entirely and probably reside in train.cpp.