Temporal-difference learning has been lifted to playernet.
The neural-net proper no longer knows anything about temporal-difference
learning, and this is instead implemented by Learner, a specialization of
NeuralNetPlayer.

Any model that wants to use that facility simply needs to export the
model Parameters to a class and provide a gradient calculator with
the required simple algebraic operators.