Ensure total symmetry between black and white at game startup.
In edge cases, there was an asymmetry between black and white at
game startup, where white and black would receive their initial
dice ordered differently.  This is of no importance except in
edge cases such as when all the equities are the same for all
the candidates.  In that case, black and white will choose different
moves because the order of enumeration is different.

This fix makes black and white receive their dice in the same
order whenever they win the opening roll, namely the high
die first and the low die second.  So there should now be absolute
symmetry between black and white with respect to play.