Use -march only for gcc.
The original matrix multiply seems at least a little faster than
both of the alternatives shown.

We learned, however, that the -march specification is very useful
for gcc++-10 here, but makes clang++-10 uncompetitive.  Without
the -march switch, clang++-10 is competetive, but a little slower,
at least on this Thinkpad T430.  About five percent.