Use gradient mean in batch learning.
In batch learning, the gradient that we apply for the update is now
divided by the batch size, consistent with ordinary practice.

Previously, we simply accumulated the sum of the gradients in the
batch.  According to the new and conventional way, the effect was
to scale the learning rate with the batch size.

So now alpha is the the learning rate, and is applied to the gradient
or the mean gradient (for batch size greater than one).  We can choose
how to adjust alpha ourselves as the batch size changes.

Note that this is implemented by simply scaling the alpha in the
LearningOptions that TemporalDifference uses.